split context files into folders
