+++ src/core/context/consume/context-consumer.mixin.ts
@@ -1,4 +1,4 @@
-import type { HTMLElementConstructor } from '../models';
+import type { HTMLElementConstructor } from '../../models';
 import { UmbContextConsumer } from './context-consumer';
 
 export declare class UmbContextConsumerInterface {
