+++ src/core/context/provide/context-provider.test.ts
@@ -1,7 +1,7 @@
 import { expect } from '@open-wc/testing';
+import { UmbContextConsumer } from '../consume/context-consumer';
+import { UmbContextRequestEventImplementation } from '../consume/context-request.event';
 import { UmbContextProvider } from './context-provider';
-import { UmbContextConsumer } from './context-consumer';
-import { UmbContextRequestEventImplementation } from './context-request.event';
 
 class MyClass {
 	prop = 'value from provider';
